+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+Version 1.7.7 (26-06-09)
+------------------------
+* new features for SPTK: lfcc and cqcc
+* configurable audMLD feature sets
+* handle multiple name-target combinations in the flags module
+* handle also explore module with flags, beneach nkululeko
+
+
 Version 1.7.6 (26-06-03)
 ------------------------
 * ease the transfer of projects by aligning file_paths